Move advanced settings into its own struct.
