+++ Scripts/Runtime/Core/CollectionAdvancedSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4c64aa67efe3f44b497b97200eedcb06
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
